[optimizer] move OnlineOptimizerManager under manager/online_runtime
