+++ kv_cache_manager/optimizer/test/online_optimizer_manager_test.cc
@@ -3,7 +3,7 @@
 
 #include "kv_cache_manager/common/unittest.h"
 #include "kv_cache_manager/optimizer/config/optimizer_registry_manager.h"
-#include "kv_cache_manager/optimizer/online_runtime/online_optimizer_manager.h"
+#include "kv_cache_manager/optimizer/manager/online_runtime/online_optimizer_manager.h"
 
 namespace kv_cache_manager {
